Reorganize file structure
- Move files into `lib`
  - Keep jobs in `lib/jobs`

- Split up Rakefile
  - Move rake tasks into `lib/tasks/**/*.rake`
  - Move configurations into `config/**/*.rb`
